third_party/blink/web_tests/external/wpt/inert/inert-inlines-around-selection-range-in-contenteditable-expected.txt

This is a testharness.js-based test.
[FAIL] <span inert style="display:none">a[bc</span><span>de]f</span>
  assert_equals: execCommand("delete") at <span inert style="display:none">a[bc</span><span>de]f</span>: <span> content should not be deleted because anchor node of Selection is in the <span inert> expected "<span inert=\\"\\" style=\\"display:none\\">abc</span><span>def</span>" but got "<span>f</span>"
[FAIL] <span inert style="display:none">{abc</span><span>de]f</span>
  assert_equals: execCommand("delete") at <span inert style="display:none">{abc</span><span>de]f</span>: <span> content should not be deleted because anchor node of Selection is in the <span inert> expected "<span inert=\\"\\" style=\\"display:none\\">abc</span><span>def</span>" but got "<span>f</span>"
Harness: the test ran to completion.
